feat: search movies by category and add update to all controllers  endpoints

+++ src/main/java/dev/nickzs/moviefliex/controller/CategoryController.java
@@ -1,9 +1,12 @@
 package dev.nickzs.moviefliex.controller;
 
 import dev.nickzs.moviefliex.controller.request.CategoryRequest;
+import dev.nickzs.moviefliex.controller.request.MovieRequest;
 import dev.nickzs.moviefliex.controller.response.CategoryResponse;
+import dev.nickzs.moviefliex.controller.response.MovieResponse;
 import dev.nickzs.moviefliex.entity.Category;
 import dev.nickzs.moviefliex.mapper.CategoryMapper;
+import dev.nickzs.moviefliex.mapper.MovieMapper;
 import dev.nickzs.moviefliex.service.CategoryService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -11,6 +14,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/moviefliex/category")
@@ -28,27 +32,38 @@ public class CategoryController {
         return ResponseEntity.ok(categories);
     }
 
-
     @PostMapping()
     public ResponseEntity<CategoryResponse> saveCategory(@RequestBody CategoryRequest request) {
         Category newCategory = CategoryMapper.toCategory(request);
         newCategory = categoryService.save(newCategory);
 
-        return ResponseEntity.status(HttpStatus.CREATED).body(
-                CategoryMapper.toCategoryResponse(newCategory)
-        );
+        return ResponseEntity.status(HttpStatus.CREATED)
+                .body(CategoryMapper.toCategoryResponse(newCategory));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<CategoryResponse> findById(@PathVariable long id) {
+    public ResponseEntity<CategoryResponse> findById(@PathVariable Long id) {
         return categoryService.findById(id)
                 .map(category -> ResponseEntity.ok(CategoryMapper.toCategoryResponse(category)))
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable long id){
-        categoryService.deleteById(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    public ResponseEntity<Void> deleteById(@PathVariable Long id) {
+        Optional<Category> optionalCategory = categoryService.findById(id);
+
+        if (optionalCategory.isPresent()) {
+            categoryService.deleteById(id);
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        }
+
+        return ResponseEntity.notFound().build();
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<CategoryResponse> update(@PathVariable Long id, @RequestBody CategoryRequest request) {
+        return categoryService.update(id, CategoryMapper.toCategory(request))
+                .map(category -> ResponseEntity.ok(CategoryMapper.toCategoryResponse(category)))
+                .orElse(ResponseEntity.notFound().build());
     }
 }
