+++ src/main/java/dev/nickzs/moviefliex/controller/MovieController.java
@@ -11,6 +11,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/moviefliex/movie")
@@ -41,15 +42,38 @@ public class MovieController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<MovieResponse> findById(@PathVariable long id) {
+    public ResponseEntity<MovieResponse> findById(@PathVariable Long id) {
         return movieService.findById(id)
                 .map(streaming -> ResponseEntity.ok(MovieMapper.toMovieResponse(streaming)))
                 .orElse(ResponseEntity.notFound().build());
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable long id){
-        movieService.deleteById(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    public ResponseEntity<Void> deleteById(@PathVariable Long id){
+        Optional<Movie> optMovie = movieService.findById(id);
+
+        if(optMovie.isPresent()) {
+            movieService.deleteById(id);
+
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        }
+        return ResponseEntity.notFound().build();
     }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<MovieResponse> update(@PathVariable Long id, @RequestBody MovieRequest request) {
+        return movieService.update(id, MovieMapper.toMovie(request))
+                .map(movie -> ResponseEntity.ok(MovieMapper.toMovieResponse(movie)))
+                .orElse(ResponseEntity.notFound().build());
+    }
+
+    @GetMapping("/search")
+    public ResponseEntity<List<MovieResponse>> findByCategory(@RequestParam Long categoryId) {
+        List<MovieResponse> movies = movieService.findByCategory(categoryId).stream()
+                .map(MovieMapper::toMovieResponse)
+                .toList();
+
+        return ResponseEntity.ok(movies);
+    }
+
 }
