+++ src/main/java/dev/nickzs/moviefliex/controller/StreamingController.java
@@ -1,8 +1,12 @@
 package dev.nickzs.moviefliex.controller;
 
+import dev.nickzs.moviefliex.controller.request.CategoryRequest;
 import dev.nickzs.moviefliex.controller.request.StreamingRequest;
+import dev.nickzs.moviefliex.controller.response.CategoryResponse;
 import dev.nickzs.moviefliex.controller.response.StreamingResponse;
+import dev.nickzs.moviefliex.entity.Category;
 import dev.nickzs.moviefliex.entity.Streaming;
+import dev.nickzs.moviefliex.mapper.CategoryMapper;
 import dev.nickzs.moviefliex.mapper.StreamingMapper;
 import dev.nickzs.moviefliex.service.StreamingService;
 import lombok.RequiredArgsConstructor;
@@ -11,6 +15,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Optional;
 
 @RestController
 @RequestMapping("/moviefliex/streaming")
@@ -47,8 +52,23 @@ public class StreamingController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteById(@PathVariable long id){
-        streamingService.deleteById(id);
-        return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+    public ResponseEntity<Void> deleteById(@PathVariable Long id) {
+        Optional<Streaming> optionalStreaming = streamingService.findById(id);
+
+        if (optionalStreaming.isPresent()) {
+            streamingService.deleteById(id);
+            return ResponseEntity.status(HttpStatus.NO_CONTENT).build();
+        }
+
+        return ResponseEntity.notFound().build();
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<StreamingResponse> update(@PathVariable Long id, @RequestBody StreamingRequest request) {
+        return streamingService.update(id, StreamingMapper.toStreaming(request))
+                .map(streaming -> ResponseEntity.ok(StreamingMapper.toStreamingResponse(streaming)))
+                .orElse(ResponseEntity.notFound().build());
+
+
     }
 }
